Add blog and RSS feed function, switch to pipenv for package management"
This commit introduces several key updates to the `mkdocs.yml` file, `Dockerfile` and `dev.Dockerfile`. Importantly, a blog and RSS feed functionality have been added to the project. Specifically, configuration for a new blog and RSS feed is included in the `mkdocs.yml` file, allowing for the creation and management of website blogs and creation of an RSS feed for blog posts.

Also, this commit changes the package management for the project Dockerfiles from pip to pipenv. This provides better management of dependencies. This is seen in both the `Dockerfile` and `dev.Dockerfile`, where the old approach of installing packages individually with pip is replaced by the installation of pipenv and subsequently running pipenv install. This is expected to streamline package management and reduce potential issues with package versions and dependencies.