Update "more" tag and revise link paths in blog posts
This commit removes the "more" tag delimiter from the mkdocs configuration and updates the German version link in the blog post. The mkdocs.yml configuration file no longer needs the "more" tag delimiter - the document displays in full. This makes the reading experience smoother and more enjoyable for our users. The blog posts' link to the German version has been updated following a change in the file path structure, ensuring that links are always directing to the correct places.