Add English version link to German Minecraft blog post
This commit inserts a link to the English version of the Minecraft main class blog post in the corresponding German blog post. This allows the German readers to easily access the English version of the post if they want to, improving the user experience on our site. Additionally, any ambiguities or confusion due to the language