Add robots.txt file and update mkdocs.yml
A robots.txt file was added to allow all data to be indexed by web robots, helping improve SEO. A sitemap.xml was also indicated for better web crawling. In the mkdocs.yml file, logo and favicon were introduced for better branding and the dark mode toggle icon was updated to a more intuitive one, enhancing the user experience.