Add new blog posts about redirecting with Traefik
Added two blog posts in English and German respectively, describing the process to set up redirection to an external URL using Traefik in a Docker environment. Also included a shell script which provides an automated process of creating redirects. These posts aim to provide clear instructions and automation for users needing to set redirections with Traefik.