Correct capitalization and remove redundant content
This commit includes a series of corrections to the blog post on "Gradle Minecraft Basic and Advanced". The changes are majorly related to the capitalization of proper nouns such as Paper, Spigot, Maven, Gradle etc. which were previously incorrectly written in lowercase. Additionally, it eliminates an unnecessary section heading towards the middle of the article, enhancing the flow of reading. The commit also makes several other corrections and additions for semantic and technical accuracy, especially related to Paperweight and Maven Central usage and explanations. It also includes translation of the same changes into German for the corresponding German blog post.