Update favicon path in mkdocs.yml
The favicon path in the mkdocs.yml file is updated from "assets/images/favicon.png" to "assets/favicon.png". This change is necessary to reflect the new location of the favicon file within the assets directory structure. No functionality is affected by this change, it just stays in line with the updated project structure.