Revise Gradle dependency documentation
The documentation sections regarding `compileOnly` dependencies in both the English and German versions of the Gradle Minecraft tutorial were revised. The previous wording suggested that `compileOnly` only applies to Paper, which was misleading. It's been clarified that any plugin APIs, not just Paper, should be imported as `compileOnly`. Moreover, a new statement about APIs not available as plugins, like Inventory APIs, and how they should be imported as `implementation` was added. This is in line with widespread practices and provides clearer guidance for users.